cf-solana: add bank hash handling plus some refactoring (#370)
Firstly, add DeltaHashProof type which contains all the information that
are needed to calculate delta hash. One piece of information in it is
the accounts delta hash. As such, this struct together with a trusted
bank hash acts as a proof for the accounts delta hash.

Secondly, create AccountHashData wrapper for the account_hash_data field
of the AccountProof struct. This improves type safety and better
encapsulates the account information used in hash calculation.

Lastly, fix a few instances where solana-program2 Cargo feature name was
used instead of proper solana-program-2.