fast-bridge: Send required accounts for cross domain transfer (#375)
For cross domain transfers, an acknowledgement is sent from destination
to source in form of a token transfer. The packet contains the memo that
has the information about unlocking the funds. The source would then
parse the memo and unlock the funds to the solver mentioned in the memo.
Since we only have token transfers enabled and no cross chain messages,
a token needs to be transferred. For this, we create a dummy token owned
by the program which mints a token everytime and sends a transfer. The
timeout of the transfer is set to infinite so that it always reaches the
counterparty chain.

Also the `on_receive_transfer` method accepts a memo which unlocks the
funds to the solver. Right now it is being called by the auctioneer but
once the hooks are enabled on `solana-ibc` program, this method would
then be called by the `solana-ibc` program.