solana-ibc: replace restaking staking program id with bridge contract program id (#382)
When inner CPI calls are being made, the called program can only know
the initial program which made the CPI. It cannot know if there were
inner CPI calls being made before calling the last one. So to whitelist
the `update_stake` method, it should check if the caller is
bridge-contract program ID instead of restaking program ID since its the
bridge contract that would call the restaking and then the solana-ibc.