cf-solana: fix unused imports warning when importing lib::par::prelude
Depending whether lib crate is built with Rayon support or not, the
lib::par::prelude module is populated by Rayon traits or empty.
lib::par clients use the prelude to be able to call Rayon trait
methods on Chunks object returned by chunks function.

Alas, when Rayon is not enabled, the prelude is empty and the compiler
recognises it as an unused import.  Disable the warning compiler
produces.