solana-ibc: get host consensus state based on feature (#401)
Host consensus state is required for establishing the connection. So the
host consensus state should be returned based on whether the `witness`
feature is present or not. If the feature is present, then we return the
rollup consensus state else we return the guest consensus state. But the
issue is if the `witness` feature is not enabled, it could either be
guest or wasm ( which is a wrapper version of guest ). Since the
connection with cosmos is already established, we would stick to
returning the `guest` consensus state for now.