IBC Hooks: add hooks to call bridge escrow program (#403)
A message is sent through IBC to unlock the funds of the solver. To
achieve this, the solana-ibc program needs to forward the memo to the
`bridge-escrow` program so that it can release the funds to the solver.
Since anybody can send memo, we execute the hook only if the transferred
token is the token owned by the `bridge-escrow` on the counterparty
chain (ethereum). This allows us to verify that the message to unlock
the funds originated from the right contract and it is not spoofed.

---------

Co-authored-by: Michal Nazarewicz <mina86@mina86.com>