hooks: fix hook structure by deserializing json instead of string (#404)
Since PFM only support passing of the memo for the next transfer as
json, strings can be passed. So we have to deserialize the memo with
json and get the string from it. Also log the acknowledgement after the
hooks is performed since the ack can change based on the result of
hooks.